+++ src/types/file/browserFile.ts
